+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### WIP
 
 - Remove bg on quoted strings inside metadata
+- Handle eval of `#_` forms in nREPL JVM
 
 ### 4.2.2 - Sep 27, 2024